+++ lib/src/theme.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 const double _iconSize = 16.0;
+const double _dialogBorderRadius = 8.0;
 const String dnbFont = 'DNB';
 final Color _statusbarDarkColor = Colors.transparent;
 final Color _statusbarLightColor = Colors.transparent;
@@ -29,6 +30,12 @@ class Eufemia {
       fontFamily: dnbFont,
       textTheme: lightTextTheme,
       bottomAppBarColor: Colors.white,
+      toggleableActiveColor: EufemiaColors.seaGreenAlt,
+      dialogTheme: DialogTheme(
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(_dialogBorderRadius),
+        ),
+      ),
     );
   }
 
@@ -50,6 +57,12 @@ class Eufemia {
       ),
       textTheme: darkTextTheme,
       bottomAppBarColor: Colors.black,
+      toggleableActiveColor: EufemiaColors.seaGreenAlt,
+      dialogTheme: DialogTheme(
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(_dialogBorderRadius),
+        ),
+      ),
     );
   }
 
@@ -84,12 +97,10 @@ class Eufemia {
   ///Large button
   static TextStyle get buttonLarge => TextStyle(
         fontSize: 16.0,
-        height: 1.33,
       );
 
   static TextStyle get buttonSmall => TextStyle(
         fontSize: 14.0,
-        height: 1.33,
         fontWeight: FontWeight.w500,
       );
 
@@ -97,53 +108,53 @@ class Eufemia {
   static TextStyle get titleLarge => TextStyle(
         fontWeight: FontWeight.bold,
         fontSize: 34,
-        height: 1.33,
+        height: 1.4,
       );
 
   /// 24pt Regular
   static TextStyle get titleMedium => TextStyle(
         fontSize: 24,
-        height: 1.33,
+        height: 1.4,
       );
 
   /// 24pt Bold
   static TextStyle get titleMediumEmphasized => TextStyle(
         fontWeight: FontWeight.bold,
         fontSize: 24,
-        height: 1.33,
+        height: 1.4,
       );
 
   /// 20pt Regular
   static TextStyle get title => TextStyle(
         fontSize: 20,
-        height: 1.33,
+        height: 1.4,
       );
 
   /// 20pt Semibold
   static TextStyle get titleDemi => TextStyle(
         fontWeight: FontWeight.w500,
         fontSize: 20,
-        height: 1.33,
+        height: 1.4,
       );
 
   /// 20pt Bold
   static TextStyle get titleEmphasized => TextStyle(
         fontWeight: FontWeight.bold,
         fontSize: 20,
-        height: 1.33,
+        height: 1.4,
       );
 
   ///  17pt Regular
   static TextStyle get body => TextStyle(
         fontSize: 17,
-        height: 1.33,
+        height: 1.4,
       );
 
   /// 17pt Bold
   static TextStyle get bodyEmphasized => TextStyle(
         fontWeight: FontWeight.bold,
         fontSize: 17,
-        height: 1.33,
+        height: 1.4,
       );
 
   /// 15pt Regular